requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
reportlab==4.0.4
python-dateutil==2.8.2
Werkzeug==2.3.7
python-dotenv==1.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
